streamer/requirements.txt
numpy==2.1.0
opencv-python==4.10.0.84
pillow==10.4.0
